[analyzer] Reorder the layout of `MemRegion` and cache by hand for optimal size
By the time working on the gdb pretty-printers for the Analyzer classes,
I observed that the `MemRegion` class has an inefficient memory layout.

Each and every object which's has a `MemRegion` as a subobject would benefit from
this change. We could spare 8 bytes for each case.
As a downside, I had to substitute the `llvm::Optional`, since that alone
consumes 16 bytes (even if the wrapped object would consume 8 bytes).

Before the patch: `sizeof(MemRegion) == 48`

Reviewed By: ASDenysPetrov

Differential Revision: https://reviews.llvm.org/D86295