[clang-tidy] Improve cert-dcl58-cpp check
The primary motivator for this patch was the second example of:
https://wiki.sei.cmu.edu/confluence/display/cplusplus/DCL58-CPP.+Do+not+modify+the+standard+namespaces

The example was about an explicit specialization, which was outright
ignored by the previous implementation.

After this patch, a warning will be emitted the specializations which
only refer to //standard// or //builtin// types.

Specializations, refering to //program-defined// types would be still
ignored, as they are allowed by the standard given that the semantics of
the specialization matches the generic version.

Template types and template type packs are also considered in the
specializations. There are a couple things which needs to be done prior
submitting this patch:
 - Refactor
 - Define and use helper AST-matchers to hoist the matching and
   suppression logic out of the business-logic.
 - There are 4 FIXMEs in the `dcl58-cpp.cpp` test which needs to be
   resolved.
 - Mention this in the release notes, also update the checker's doc
   page.

So far it did not crash, nor introduce performance regression.